rename store to +store
