+++ src/app/+store/app.model.ts
@@ -1,4 +1,8 @@
+import { Scenario, TrackFile } from "../types/data.types";
+
 export interface AppStore {
+  scenarios: Scenario[],
+  trackFiles: TrackFile[]
   satellites: Satellite[];
   selectedSatId: number | null;
 }
