+++ src/app/+store/app.reducer.ts
@@ -1,11 +1,16 @@
 import { createFeature, createReducer, on } from '@ngrx/store';
 import { SatelliteActions } from './app.actions';
 import { AppStore } from './app.model';
+import { mockScenarios, mockTrackFiles } from '../mock-data/generate-data';
 
 export const initialState: AppStore = {
+  scenarios: mockScenarios,
+  trackFiles: mockTrackFiles,
   satellites: [{ catalogId: 123 }],
   selectedSatId: null,
 };
+
+console.log(initialState.trackFiles)
 
 export const SatellitesReducer = createReducer(
   initialState,
