+++ src/app/main/main.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core'
 import { Store } from '@ngrx/store'
-import { selectSatellites, selectSelectedSatId } from '../store/app.reducer'
-import { selectSelectedSatellite } from '../store/app.selectors'
+import { selectSatellites, selectSelectedSatId } from '../+store/app.reducer'
+import { selectSelectedSatellite } from '../+store/app.selectors'
 import { AstroComponentsModule } from '@astrouxds/angular'
 import { UtilityToolkitComponent } from './utility-toolkit/utility-toolkit.component'
 import { CommonModule } from '@angular/common'
