+++ src/app/scenario-library/scenario-library.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, signal, ViewChild } from '@angular/core';
 import { AstroComponentsModule, RuxToastStack } from '@astrouxds/angular';
 import { Store } from '@ngrx/store';
-import { SatelliteActions } from '../store/app.actions';
+import { SatelliteActions } from '../+store/app.actions';
 import { ToastService } from '../shared/toast.service';
 @Component({
   standalone: true,
