+++ src/main.ts
@@ -3,7 +3,7 @@ import { AppComponent } from './app/app.component';
 import { provideRouter } from '@angular/router';
 import { Routes } from './app/app-routing';
 import { provideState, provideStore } from '@ngrx/store';
-import { satellitesFeature } from './app/store/app.reducer';
+import { satellitesFeature } from './app/+store/app.reducer';
 
 bootstrapApplication(AppComponent, {
   providers: [
